Added Axes, Delete Model and onFrame on Setup
Added drawAxes method to the scene to automatically draw the axes in the render.

Added a new method to the scene to delete a model from the scene.

Added two functions to the code that will be called, one at the beggining of the rendering and another that will be called in every frame.